separated controller functions to their own files
