+++ .gitignore
@@ -25,3 +25,4 @@ dist-ssr
 .env
 ./src/firebase/serviceAccountKey.json
 server/pnpm-lock.yaml
+server/assets/noises
