+++ server/src/controllers/status.ts
@@ -1,0 +1,7 @@
+import { Request, Response } from 'express'
+
+export const checkSanity = (req: Request, res: Response) => {
+  res.status(200).json({
+    status: 'on',
+  })
+}
